Controller subscribe action that dispatched only main process.
## About

Previous Controller is located in IPC communication logic,
So controller subscribe action about only renderer process to main process.

It is improved that controller subscribe every action whaterver main to main or renderer to renderer.

## Issue
When improve Controller, We have global import module issue.
Some local module like `components/i18n/index.js` and `store/createStore` is using both process that main and renderer.
It doesn't matter because we write a code conditionally if it is used in main process or renderer process.
But globally `import` make error. For example, `store/storeMain.js` must use only main process. But It imported in renderer process although not using.
That make error, not using but import!,
So we solve that conditionally import using `require` in ES5.